Fully remove active store from engine.
